Adds check for optional deps in INSTALLED_APPS
Previously the serializers.py file relied solely on the presence of
allauth.socialaccount in the PYTHON_PATH to determine if its use was
required. This adds another check in the Django INSTALLED_APPS for the
allauth.socialaccount app, and then continues with the import if the app
has been added.